Lirc improvement (#8)
* Respect virtualenv interpreter

* Add script to create lircrc config file from lircd.conf

* Reformat

* Respect remote value if given from lirc and use in MQTT.

This checks if the remote name was given from lirc (as done with lircrc
files generated by create_lircrc.py). If so, the remote will be added to
the mqtt topic path.

* Extend README

* Rename example config

* Fix unexpected behaviour

The doc of the python lirc module said, lists are given as list. In fact
the config comes as single string.

* Upgrade phao

* Update readme

* Set maxsplit for remote,key split

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>

* Extend example lirc conf with remote name.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>

* Make create_lirc.py executeable

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>

* Extend readme with lirc details

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>